% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_by_n.R
\name{add_by_n}
\alias{add_by_n}
\title{bl_nobs_function}
\usage{
add_by_n(data, variable, by, ...)
}
\arguments{
\item{data}{}

\item{variable}{}

\item{by}{}

\item{...}{}
}
\description{
This function can be used to calculate the number of Nobs (i.e. number of non-missing data).
It can be used to add the Nobs to a new column into a gtsummary table (see example).
}
\examples{
 table \%>\%
 add_stat( fns = everything() ~ add_by_n    )
 (where table is a gtsummary table)
 
}
