% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nobs.R
\name{add_nobs}
\alias{add_nobs}
\title{bl_nobs_function}
\usage{
add_nobs(table, location = "col")
}
\arguments{
\item{table}{a gtsummary table of class "tbl_summary"}

\item{location}{the location for the Nobs.Can take the following values:
 
"col" = Nobs is displayed new column(s) (default)  

"row" : Nobs is displayed as statistics alongside the variable name. Stats are presented below 
 
"var_name": Nobs is added after the variable name within the same cell.This option is only available if the table is not stratified}
}
\value{
a gtsummary table with the Nobs
}
\description{
Add Nobs to a gtsummary table of class "tbl_summary" 
This function add the Nobs (i.e. number of non-missing data) to a gtsummary table. 
Depending on the specified location Nobs will be:
(i) added in a new column, 
(ii) displayed as statistics alongside the variable name
(iii) added after the variable name within the same cell.
}
\examples{
#Data
library(dplyr)
library(labelled)
library(gtsummary)

data("cancer", package = "survival")#' 
cancer <- 
 cancer |> 
mutate(
  sex.factor = factor(sex, levels = c(1,2), labels = c("Female", "Male") ),
 yn_var = factor(sex, levels = c(1,2), labels = c("Yes", "No") )
) |>
set_variable_labels (
 yn_var = "Yes/no variable",
 sex.factor = "Gender",
 inst = "Institution code",
  time = "Time", 
 age = "Age",
 ph.ecog = "ECOG score") 
 

 # Ex1: stratified table
 #. gtsummaty table
tab <- 
  cancer |> 
  select(sex.factor, yn_var, inst, time,age, ph.ecog) |> 
  tbl_summary(
 by=sex.factor,
 type = all_continuous() ~ "continuous2",
 statistic = all_continuous() ~ c("{median} ({p25}, {p75})", "{min} - {max}"),
)  |> 
add_overall() 

 # .Add Nobs gtsummaty table 
tab |>  add_nobs (location = "row")
tab |> add_nobs (location = "col")
  
# Ex2: un-stratified table 
#. gtsummary table
tab <- 
  cancer |> 
  select(sex.factor, yn_var, inst, time,age, ph.ecog) |> 
  tbl_summary() 
  
 # .Add Nobs gtsummaty table 
 tab |>  
 add_nobs(location = "var_name")

}
